requirements.txt
numpy==2.3.2
Pillow==11.3.0
pydicom==3.0.1